Move WriteThread::Writer to separate file (pipeline write part 2)
Summary: Yet another refactor to make it easier for me to plugin the
pipeline write code. Moving `WriteThread::Writer` to its separate file and
move state transition methods (`BlockingAwaitState()`, `AwaitState()` and
`SetState()`) to `Writer`. Later I'll reuse the `Writer` class in
pipeline write implementation. I'm also changing the class with the same
name in block_cache_tier_file.h to `FileWriter` since there's naming
conflict.

Stacks on #2042.

Test Plan:
Since there's no logic change, I'll just make sure existing test all
pass.
