+++ db/db_impl_debug.cc
@@ -127,13 +127,13 @@ void DBImpl::TEST_UnlockMutex() {
 }
 
 void* DBImpl::TEST_BeginWrite() {
-  auto w = new WriteThread::Writer();
+  auto w = new Writer();
   write_thread_.EnterUnbatched(w, &mutex_);
   return reinterpret_cast<void*>(w);
 }
 
 void DBImpl::TEST_EndWrite(void* w) {
-  auto writer = reinterpret_cast<WriteThread::Writer*>(w);
+  auto writer = reinterpret_cast<Writer*>(w);
   write_thread_.ExitUnbatched(writer);
   delete writer;
 }
